+++ apis/v1/root.api.js
@@ -1,7 +1,7 @@
 const router = require("express").Router();
-const commonEndpoints = require("./routes/common.api")
-const webEndpoints = require("./routes/web.api")
-const adminEndpoints = require("./routes/admin.api")
+const commonEndpoints = require("./routes/common_api")
+const webEndpoints = require("./routes/web_api")
+const adminEndpoints = require("./routes/admin_api")
 
 router.use("/common", commonEndpoints);
 router.use("/web", webEndpoints);
