+++ configs/parish_schema.json
